Moves the --wait test to a different file
